requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
hydra-core>=1.3.0
ucimlrepo>=0.0.3
optuna>=3.3.0 